tools: Move silencio to logger.py
