+++ devito/__init__.py
@@ -11,7 +11,7 @@ from devito.dimension import *  # noqa
 from devito.equation import *  # noqa
 from devito.finite_difference import *  # noqa
 from devito.grid import *  # noqa
-from devito.logger import error, warning, info, set_log_level  # noqa
+from devito.logger import error, warning, info, set_log_level, silencio  # noqa
 from devito.parameters import *  # noqa
 from devito.tools import *  # noqa
